Delete existing content
I initially meant to use this site for blogging, but I did not stick with it and it quickly devolved to negativity spewing. I am thinking of blogging on hubelbauer.net now and am not sure what I'll do with this repository now, but I know this content can go and I might drop it from the Git history later, too.