Copy over and adjust the relevant hubelbauer.net pipelines
I have developed a good suite of checks for hubelbauer.net and many will apply here as well, so here they are copy-pasted and adjusted to use the correct domain name.

Many of these won't work out of the gate (like the SSL certificate, async-await.cz redirect etc.) but will already be useful as indicators of what I need to finish.