Adjust the SSL certificate to use the GitHub agent output
As the comment at the top of the file states, the agent sees consistent output, but different from the output my Mac sees, which is also consistent on it. I am replacing the output from my Mac with the output from the agent and as a result this pipeline will being to work and monitor for changes as the agent sees them.