Remove the www+apex custom domain setup task
To my surprise, it seems like I have already set this up before! That's cool, because I thought I have done it for the first time for hubelbauer.net, but no, checking in the browser even including the redirects, it just works.

For good measure, I entered the Pages section of the repository Settings page and it did not start provisioning a new certificate, so the current one is already correct. I also untoggled and toggled again the Enforce HTTP checkbox and it worked fine without kicking off certificate provisioning.

So as far as I can see, no additional work is needed here.